requirements.txt
Django==2.1.5
django-ckeditor==5.4.0
django-cleanup==2.1.0
django-crispy-forms==1.7.2
django-import-export==2.0.1
django-js-asset==1.0.0
djangorestframework==3.11.0
flake8==3.5.0
Flask==0.12.2
numpy==1.16.2
oauth==1.0.1
pandas==0.22.0
pipenv==2018.11.26
simplejson==3.13.2
sqlparse==0.3.0
virtualenv==16.7.9
